% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportRSA.fit.R
\name{exportRSA.fit}
\alias{exportRSA.fit}
\title{Export fit indices and names of polynomials models}
\usage{
exportRSA.fit(RSAcompare_object)
}
\arguments{
\item{RSAcompare_object}{A list generated by \code{RSAcompare(RSA_object)}}
}
\value{
A table of fit indices and names of polynomial families or user model
}
\description{
Utilitary function to easily export fit indices of polynomial models (predefined or user-specific), including their proper theoretical names (for predefined polynomial families) and the fit indices used for comparison.
}
