% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RSA_step1.R
\docType{data}
\name{RSA_step1}
\alias{RSA_step1}
\title{Simulation data on needs-supplies fit processes (STEP1)}
\description{
An object of class "RSA" containing estimates of all polynomial families, 
obtained by passing the function \code{RSAmodel(formula= engagement ~ needs*supplies,
data= sim_NSfit, model= c("CUBIC","STEP1"))}. Used to illustrate functions.
}
\author{
Fernando Núñez-Regueiro 
\email{fernando.nr.france@gmail.com}
}
\keyword{data}
